move all the added tests into test_timeline_delete.py
The aspect of remote storage needs to be considered in each test.
It's nonsense to have them all in test_remote_storage.py.

Test test_timeline_deletion_with_files_stuck_in_upload_queue should
probably be moved as well, but, it existed before this PR.
